+++ lib/src/framework.dart
@@ -6,6 +6,7 @@ import 'package:flutter/scheduler.dart';
 import 'package:flutter/widgets.dart';
 import 'package:provide_it/src/injector/injector.dart';
 import 'package:provide_it/src/injector/param.dart';
+import 'package:provide_it/src/utils/async_snapshot_extension.dart';
 import 'package:provide_it/src/watchers/listenable.dart';
 
 import '../provide_it.dart';
@@ -19,3 +20,12 @@ part 'framework/provide_it_element.dart';
 part 'framework/ref_state.dart';
 part 'framework/ref_state_cycling.dart';
 part 'framework/watcher.dart';
+
+extension ProvideItExtension on BuildContext {
+  @protected
+  ProvideItElement get provideIt {
+    final it = getElementForInheritedWidgetOfExactType<ProvideIt>();
+    assert(it != null, 'You must set `ProvideIt` in your app.');
+    return it as ProvideItElement;
+  }
+}