+++ lib/src/framework/ref_state_cycling.dart
@@ -58,7 +58,7 @@ extension<T, R extends Ref<T>> on RefState<T, R> {
       /// on reassemble, [didUpdateRef] should always be called.
       /// null implies that the ref was removed, allowing safe disposal.
       if (_lastRef == null) {
-        ProvideItElement.instance._disposeRef(context, ref);
+        _root?._disposeRef(context, ref);
       }
     });
   }